+++ src/main/java/com/ikubinfo/plumbershop/product/model/ProductDocument.java
@@ -13,8 +13,8 @@ public class ProductDocument extends BaseDocument {
 
     private String name;
     private String description;
-    private Long buyingPrice;
-    private Long sellingPrice;
+    private double buyingPrice;
+    private double sellingPrice;
     private Integer count;
     private Map<String,String> attributes;
     private String code;
